Move update coordinator into coordinator module
