+++ homeassistant/components/tplink_omada/coordinator.py
@@ -5,14 +5,19 @@ from __future__ import annotations
 import asyncio
 from datetime import timedelta
 import logging
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, NamedTuple
 
 from tplink_omada_client import OmadaSiteClient, OmadaSwitchPortDetails
 from tplink_omada_client.clients import OmadaWirelessClient
-from tplink_omada_client.devices import OmadaGateway, OmadaListDevice, OmadaSwitch
+from tplink_omada_client.devices import (
+    OmadaFirmwareUpdate,
+    OmadaGateway,
+    OmadaListDevice,
+    OmadaSwitch,
+)
 from tplink_omada_client.exceptions import OmadaClientException
 
-from homeassistant.core import HomeAssistant
+from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
 if TYPE_CHECKING:
@@ -24,6 +29,7 @@ POLL_SWITCH_PORT = 300
 POLL_GATEWAY = 300
 POLL_CLIENTS = 300
 POLL_DEVICES = 300
+POLL_UPGRADE = 60
 
 
 class OmadaCoordinator[_T](DataUpdateCoordinator[dict[str, _T]]):
@@ -144,3 +150,68 @@ class OmadaClientsCoordinator(OmadaCoordinator[OmadaWirelessClient]):
             async for c in self.omada_client.get_connected_clients()
             if isinstance(c, OmadaWirelessClient)
         }
+
+
+class FirmwareUpdateStatus(NamedTuple):
+    """Firmware update information for Omada SDN devices."""
+
+    device: OmadaListDevice
+    firmware: OmadaFirmwareUpdate | None
+
+
+class OmadaFirmwareUpdateCoordinator(OmadaCoordinator[FirmwareUpdateStatus]):  # pylint: disable=hass-enforce-class-module
+    """Coordinator for getting details about available firmware updates for Omada devices."""
+
+    def __init__(
+        self,
+        hass: HomeAssistant,
+        config_entry: OmadaConfigEntry,
+        omada_client: OmadaSiteClient,
+        devices_coordinator: OmadaDevicesCoordinator,
+    ) -> None:
+        """Initialize my coordinator."""
+        super().__init__(
+            hass, config_entry, omada_client, "Firmware Updates", poll_delay=None
+        )
+
+        self._devices_coordinator = devices_coordinator
+        self._config_entry = config_entry
+
+        config_entry.async_on_unload(
+            devices_coordinator.async_add_listener(self._handle_devices_update)
+        )
+
+    async def _get_firmware_updates(self) -> list[FirmwareUpdateStatus]:
+        devices = self._devices_coordinator.data.values()
+
+        updates = [
+            FirmwareUpdateStatus(
+                device=d,
+                firmware=None
+                if not d.need_upgrade
+                else await self.omada_client.get_firmware_details(d),
+            )
+            for d in devices
+        ]
+
+        # During a firmware upgrade, poll device list more frequently
+        self._devices_coordinator.update_interval = timedelta(
+            seconds=(
+                POLL_UPGRADE
+                if any(u.device.fw_download for u in updates)
+                else POLL_DEVICES
+            )
+        )
+        return updates
+
+    async def poll_update(self) -> dict[str, FirmwareUpdateStatus]:
+        """Poll the state of Omada Devices firmware update availability."""
+        return {d.device.mac: d for d in await self._get_firmware_updates()}
+
+    @callback
+    def _handle_devices_update(self) -> None:
+        """Handle updated data from the devices coordinator."""
+        # Trigger a refresh of our data, based on the updated device list
+        self._config_entry.async_create_background_task(
+            self.hass, self.async_request_refresh(), "Omada Firmware Update Refresh"
+        )
