+++ homeassistant/components/tplink_omada/quality_scale.yaml
@@ -5,9 +5,7 @@ rules:
     status: done
     comment: Service data APIs are polled every 5 minutes
   brands: done
-  common-modules:
-    status: todo
-    comment: The coordinator for the update platform should be moved to common module.
+  common-modules: done
   config-flow-test-coverage:
     status: todo
     comment: "test_form_single_site is patching config flow internals, and should only patch external APIs. Must address feedback from #156697."
